requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==0.12

# Flask Packages
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
git+https://github.com/lingthio/Flask-User.git@master#egg=Flask-User

# Automated tests
pytest==3.0.5
pytest-cov==2.4.0
# Development tools
# Fabric3==1.13.1.post1
# tox==2.7.0

# New versions are incompatable with this flask 0.12 :(
werkzeug<1.0

